+++ internal/regopolicyinterpreter/regopolicyinterpreter_test.go
