+++ pkg/securitypolicy/securitypolicyenforcer_rego.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/Microsoft/hcsshim/internal/guest/spec"
 	"github.com/Microsoft/hcsshim/internal/guestpath"
-	rpi "github.com/Microsoft/hcsshim/pkg/regopolicyinterpreter"
+	rpi "github.com/Microsoft/hcsshim/internal/regopolicyinterpreter"
 	oci "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
 )
